Create README.md
I will be depositing in github some  Octave (~~Matlab) codes which are used to create the Figures 5 6 7 8 9 10 in my paper "A Backprojection Kernel (KRNL3D) for Very-Wide-Aperture 3D Tomography Applied to PET for Precise Use of Time-of Flight Data"  submitted to Physics in Medicine and Biology.  Also included is a "guidetocodesAug_21".  Feel free to contact me if you are serious about trying out or altering these codes and have further questions.     Keith Miller Aug10, 2021